Finishing steps
First you need to extract the ESP32 file and then flash the flash download and extract the application in it. The first file you extracted is the Router Master zip.Insert the three files in it one after the other, then press and hold the boot button on the ESP32 board until it finishes.